Added resources directory for reflex-sync
