--- src/renderer/mixins/reflex-sync/lib/eventTypes.js
@@ -1,5 +1,0 @@
-module.exports = {
-  CLICK: 'click',
-  DBLCLICK: 'dblclick',
-  SCROLL: 'scroll',
-}
