+++ src/resources/lib/effects/click.js
