+++ src/resources/lib/eventTypes.js
@@ -1,0 +1,9 @@
+const CLICK = 'click'
+const DBLCLICK = 'dblclick'
+const SCROLL = 'scroll'
+
+module.exports = {
+  CLICK,
+  DBLCLICK,
+  SCROLL,
+}
